Use size_t for GDC and get_optimal_object_size calculations as it suggested to cause no performance degradation and will always be large enough.

+++ shmem.c
@@ -46,7 +46,7 @@ static ShmSettings *shmSettings = NULL;
 static int pagesize;
 static unsigned int local_shm_counter = 0;
 
-static unsigned int get_optimal_object_size(unsigned int objsize, unsigned int minsize);
+static size_t get_optimal_object_size(size_t objsize, size_t minsize);
 
 unsigned long long addstr(const char *str)
 {
@@ -457,11 +457,11 @@ void delete_shm(SharedMemory *sharedMemory)
 }
 
 // Euclidean algorithm to return greatest common divisor of the numbers
-static unsigned int gcd(unsigned int a, unsigned int b)
+static size_t gcd(size_t a, size_t b)
 {
 	while(b != 0)
 	{
-		unsigned int temp = b;
+		size_t temp = b;
 		b = a % b;
 		a = temp;
 	}
@@ -472,14 +472,14 @@ static unsigned int gcd(unsigned int a, unsigned int b)
 // shared memory objects. This routine works by computing the LCM
 // of two numbers, the pagesize and the size of a single element
 // in the shared memory object
-static unsigned int get_optimal_object_size(unsigned int objsize, unsigned int minsize)
+static size_t get_optimal_object_size(size_t objsize, size_t minsize)
 {
-	unsigned int optsize = pagesize / gcd(pagesize, objsize);
+	size_t optsize = pagesize / gcd(pagesize, objsize);
 	if(optsize < minsize)
 	{
 		if(config.debug & DEBUG_SHMEM)
 		{
-			logg("DEBUG: LCM(%i, %u) == %u < %u",
+			logg("DEBUG: LCM(%i, %zu) == %zu < %zu",
 			     pagesize, objsize,
 			     optsize*objsize,
 			     minsize*objsize);
@@ -490,10 +490,10 @@ static unsigned int get_optimal_object_size(unsigned int objsize, unsigned int m
 		// First part: Integer division, may cause clipping, e.g., 5/3 = 1
 		// Second part: Catch a possibly happened clipping event by adding
 		//              one to the number: (5 % 3 != 0) is 1
-		unsigned int multiplier = (minsize/optsize) + ((minsize % optsize != 0) ? 1u : 0u);
+		size_t multiplier = (minsize/optsize) + ((minsize % optsize != 0) ? 1u : 0u);
 		if(config.debug & DEBUG_SHMEM)
 		{
-			logg("DEBUG: Using %u*%u == %u >= %zu",
+			logg("DEBUG: Using %zu*%zu == %zu >= %zu",
 			     multiplier, optsize*objsize,
 			     multiplier*optsize*objsize,
 			     minsize*objsize);
@@ -506,7 +506,7 @@ static unsigned int get_optimal_object_size(unsigned int objsize, unsigned int m
 	{
 		if(config.debug & DEBUG_SHMEM)
 		{
-			logg("DEBUG: LCM(%i, %u) == %u >= %u",
+			logg("DEBUG: LCM(%i, %zu) == %zu >= %zu",
 			     pagesize, objsize,
 			     optsize*objsize,
 			     minsize*objsize);
